Fix sudo usage.
Sudo is being used directly all over the place, which doesn't work when the user is already root.
Add a SUDO and SUDO_W_ENV variable that evaluates to the equivalent sudo command, but is
blanked when running as root.
Also add a sudo and sudo_w_env alias that map to a fake-root function that just runs the
command passed without any sudo call, so any attempted use of sudo by python scripts triggered from
the cmd-* scripts won't actually use sudo when running as root.